correct some types

+++ c/coins.c
@@ -3,12 +3,12 @@
 #include <time.h>
 #include <omp.h>
 
-unsigned long countSetBits(unsigned long n) {
+int countSetBits(unsigned long long n) {
   return __builtin_popcount(n);
 }
 
 // Function to count occurrences of "HH" (00) and "HT" (01)
-void countPoints(unsigned long long sequence, long *alicePoints, long *bobPoints) {
+void countPoints(unsigned long long sequence, int *alicePoints, int *bobPoints) {
     unsigned long long aliceMask = sequence & (sequence >> 1);
     unsigned long long bobMask = ~sequence & (sequence >> 1);
 
@@ -20,11 +20,11 @@ int main() {
     for (int N = 2; ; N++) {
         double start_time = omp_get_wtime();
         unsigned long long totalSequences = 1ULL << N;
-        long aliceWins = 0, bobWins = 0, draws = 0;
+        unsigned long long aliceWins = 0, bobWins = 0, draws = 0;
 
         #pragma omp parallel for reduction(+:aliceWins, bobWins, draws)
         for (unsigned long long seq = 0; seq < totalSequences; seq++) {
-            long alicePoints, bobPoints;
+            int alicePoints, bobPoints;
             countPoints(seq, &alicePoints, &bobPoints);
 
             if (alicePoints > bobPoints) {
@@ -39,7 +39,7 @@ int main() {
         double end_time = omp_get_wtime();
         double time_elapsed = end_time - start_time;
 
-        printf("n=%d, time=%f, alice wins=%lu, bob wins=%lu, draws=%lu\n", N, time_elapsed, aliceWins, bobWins, draws);
+        printf("n=%d, time=%f, alice wins=%llu, bob wins=%llu, draws=%llu\n", N, time_elapsed, aliceWins, bobWins, draws);
     }
 
     return 0;
